requirements.txt
flask-restx==0.5.1
flask-marshmallow
flask-sqlalchemy
flask-cors
gunicorn
honcho
xmltodict
marshmallow-sqlalchemy
psycopg2-binary
Flask-Migrate
Flask-Dance
Flask-Security-Too
email-validator

#dev includes
flake8
invoke
black
